pyproject.toml
[tool.poetry]
name = "conutil"
version = "1.0.0"
description = ""
authors = ["Ahmed Alam <m.ahmed.g.1982@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
